Update README.md
This is a convenient and secure **mobile non-custodial crypto wallet**, tailored for the **NEAR Protocol** ecosystem. It is well suited for storing tokens and NFTs, working with dApps and staking.
The main advantage is full control over funds (you have the keys), and the main risk is that if you lose the seed phrase, it will be impossible to restore access.
Suitable for those who already use or plan to actively use the NEAR ecosystem.